LPD-56911 request to upload must be handled by the parent, so the component MultipleFileUploader can be reused to not only upload files, but general things (like translations)
