+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main/components/modal/MultipleFilesUploadModalContent.tsx
@@ -8,10 +8,27 @@ import {openToast} from 'frontend-js-components-web';
 import {sub} from 'frontend-js-web';
 import React from 'react';
 
+import ApiHelper from '../../../services/ApiHelper';
 import {AssetLibrary} from '../../../types/AssetLibrary';
 import MultipleFileUploader, {
 	FileData,
 } from '../multiple_file_uploader/MultipleFileUploader';
+
+const getBase64 = (file: File): Promise<string> => {
+	return new Promise((resolve, reject) => {
+		const reader = new FileReader();
+		reader.onload = () => {
+			if (typeof reader.result === 'string') {
+				resolve(reader.result.split(',')[1]);
+			}
+			else {
+				reject(new Error('FileReader did not return a string.'));
+			}
+		};
+		reader.onerror = reject;
+		reader.readAsDataURL(file);
+	});
+};
 
 export default function MultipleFilesUploadModalContent({
 	assetLibraries,
@@ -30,6 +47,29 @@ export default function MultipleFilesUploadModalContent({
 }) {
 	const getAssetLibraryLink = (assetLibrary: AssetLibrary) => {
 		return `<a href="${baseAssetLibraryViewURL}${assetLibrary.groupId}" class="alert-link lead"><strong>${assetLibrary.name}</strong></a>`;
+	};
+
+	const uploadRequest = async ({
+		fileData,
+		groupId,
+	}: {
+		fileData: FileData;
+		groupId: string;
+	}) => {
+		const fileBase64 = await getBase64(fileData.file);
+
+		return await ApiHelper.post(
+			`/o/cms/basic-documents/scopes/${groupId}`,
+			{
+				file: {
+					fileBase64,
+					name: fileData.name,
+				},
+				objectEntryFolderExternalReferenceCode:
+					parentObjectEntryFolderExternalReferenceCode || 'L_FILES',
+				title: fileData.name,
+			}
+		);
 	};
 
 	const onUploadComplete = ({
@@ -91,9 +131,7 @@ export default function MultipleFilesUploadModalContent({
 				filesToUpload={filesToUpload}
 				onModalClose={onModalClose}
 				onUploadComplete={onUploadComplete}
-				parentObjectEntryFolderExternalReferenceCode={
-					parentObjectEntryFolderExternalReferenceCode
-				}
+				uploadRequest={uploadRequest}
 			/>
 		</>
 	);
