+++ modules/apps/site/site-cms-site-initializer/test/js/main/components/MultipleFileUploader.test.tsx
@@ -27,7 +27,7 @@ const DEFAULT_PROPS = {
 	],
 	onModalClose: mockCloseModal,
 	onUploadComplete: jest.fn(),
-	parentObjectEntryFolderExternalReferenceCode: 'L_FILES',
+	uploadRequest: jest.fn().mockResolvedValue({error: false}),
 };
 
 const createFile = (name: string, size: number, type = 'image/png') => {
